Changed intCodeComputer to Long.

+++ src/com/dwelguisz/year2019/ProgramAlarm.java
@@ -8,31 +8,31 @@ import java.util.List;
 public class ProgramAlarm extends AoCDay {
     public void solve() {
         List<String> lines = readFile("/Users/dwelguisz/personal/advent-of-code/src/resources/year2019/day02/input.txt");
-        Integer part1 = solutionPart1(lines, 12, 2);
-        Integer part2 = solutionPart2(lines);
+        Long part1 = solutionPart1(lines, 12L, 2L);
+        Long part2 = solutionPart2(lines);
         System.out.println(String.format("Part 1 Answer: %d",part1));
         System.out.println(String.format("Part 1 Answer: %d",part2));
     }
 
-    public Integer solutionPart1(List<String> lines, int noun, int verb) {
+    public Long solutionPart1(List<String> lines, Long noun, Long verb) {
         IntCodeComputer intCodeComputer = new IntCodeComputer();
         intCodeComputer.initializeIntCode(lines);
-        intCodeComputer.setIntCodeMemory(1, noun);
-        intCodeComputer.setIntCodeMemory(2, verb);
+        intCodeComputer.setIntCodeMemory(1L, noun);
+        intCodeComputer.setIntCodeMemory(2L, verb);
         intCodeComputer.run();
-        return intCodeComputer.getMemoryLocation(0);
+        return intCodeComputer.getMemoryLocation(0L);
     }
 
-    public Integer solutionPart2(List<String> lines) {
-        for (int noun = 0; noun < 100; noun++) {
-            for (int verb = 0; verb < 100; verb++) {
-                Integer value = solutionPart1(lines, noun, verb);
-                if (value == 19690720) {
+    public Long solutionPart2(List<String> lines) {
+        for (Long noun = 0L; noun < 100; noun++) {
+            for (Long verb = 0L; verb < 100; verb++) {
+                Long value = solutionPart1(lines, noun, verb);
+                if (value == 19690720L) {
                     return 100 * noun + verb;
                 }
             }
         }
-        return -1;
+        return -1L;
     }
